pyproject.toml
[project]
name = "modelcontextprotocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.25.0",
    "mcp>=1.10.1",
]
